+++ culturalcontent/src/main/java/com/tim5/culturalcontent/repository/NovostRepository.java
@@ -3,5 +3,5 @@ package com.tim5.culturalcontent.repository;
 import com.tim5.culturalcontent.model.Novost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NovostRepository extends JpaRepository<Novost, Long> {
+public interface NovostRepository extends JpaRepository<Novost, Integer> {
 }
