+++ culturalcontent/src/main/java/com/tim5/culturalcontent/repository/TipKPRepository.java
@@ -3,5 +3,5 @@ package com.tim5.culturalcontent.repository;
 import com.tim5.culturalcontent.model.TipKulturnePonude;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TipKPRepository extends JpaRepository<TipKulturnePonude, Long> {
+public interface TipKPRepository extends JpaRepository<TipKulturnePonude, Integer> {
 }
