+++ culturalcontent/src/main/java/com/tim5/culturalcontent/service/KulturnaPonudaService.java
@@ -11,7 +11,7 @@ public class KulturnaPonudaService implements ServiceInterface<KulturnaPonuda> {
     }
 
     @Override
-    public KulturnaPonuda findOne(Long id) {
+    public KulturnaPonuda findOne(Integer id) {
         return null;
     }
 
@@ -21,12 +21,12 @@ public class KulturnaPonudaService implements ServiceInterface<KulturnaPonuda> {
     }
 
     @Override
-    public KulturnaPonuda update(KulturnaPonuda entity, Long id) throws Exception {
+    public KulturnaPonuda update(KulturnaPonuda entity, Integer id) throws Exception {
         return null;
     }
 
     @Override
-    public void delete(Long id) throws Exception {
+    public void delete(Integer id) throws Exception {
 
     }
 }
