+++ culturalcontent/src/main/java/com/tim5/culturalcontent/service/RegistrovaniKorisnikService.java
@@ -11,7 +11,7 @@ public class RegistrovaniKorisnikService  implements ServiceInterface<Registrova
     }
 
     @Override
-    public RegistrovaniKorisnik findOne(Long id) {
+    public RegistrovaniKorisnik findOne(Integer id) {
         return null;
     }
 
@@ -21,12 +21,12 @@ public class RegistrovaniKorisnikService  implements ServiceInterface<Registrova
     }
 
     @Override
-    public RegistrovaniKorisnik update(RegistrovaniKorisnik entity, Long id) throws Exception {
+    public RegistrovaniKorisnik update(RegistrovaniKorisnik entity, Integer id) throws Exception {
         return null;
     }
 
     @Override
-    public void delete(Long id) throws Exception {
+    public void delete(Integer id) throws Exception {
 
     }
 }
