+++ culturalcontent/src/main/java/com/tim5/culturalcontent/service/TipKPService.java
@@ -11,7 +11,7 @@ public class TipKPService  implements ServiceInterface<TipKulturnePonude> {
     }
 
     @Override
-    public TipKulturnePonude findOne(Long id) {
+    public TipKulturnePonude findOne(Integer id) {
         return null;
     }
 
@@ -21,12 +21,12 @@ public class TipKPService  implements ServiceInterface<TipKulturnePonude> {
     }
 
     @Override
-    public TipKulturnePonude update(TipKulturnePonude entity, Long id) throws Exception {
+    public TipKulturnePonude update(TipKulturnePonude entity, Integer id) throws Exception {
         return null;
     }
 
     @Override
-    public void delete(Long id) throws Exception {
+    public void delete(Integer id) throws Exception {
 
     }
 }
